updated word-guessing game for modularity in future testing.
Created a few methods to handle the main events that will need to be handed. `startup()` resets the game and assigns a new random word, `take_guess()` ensures a single char is input from the user, and finally `register_guess()` completes the core logic to determine if the letter is correct and if we won/lose.